Applied review comments - split post tranche adjustment logic into its own servuce + some tidy up
