+++ src/main/kotlin/uk/gov/justice/digital/hmpps/calculatereleasedatesapi/service/timeline/BookingTimelineService.kt
@@ -12,15 +12,10 @@ import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.enumerations.Adjust
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.enumerations.AdjustmentType.RESTORATION_OF_ADDITIONAL_DAYS_AWARDED
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.enumerations.AdjustmentType.TAGGED_BAIL
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.enumerations.AdjustmentType.UNLAWFULLY_AT_LARGE
-import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.enumerations.CalculationRule
-import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.enumerations.CalculationRule.ADJUSTED_AFTER_TRANCHE_COMMENCEMENT
-import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.enumerations.ReleaseDateType
-import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.enumerations.SDSEarlyReleaseTranche
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model.Adjustments
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model.CalculableSentence
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model.CalculationOptions
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model.CalculationOutput
-import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model.CalculationResult
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model.ExternalMovement
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model.ExternalMovementDirection
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model.Offender
@@ -35,7 +30,6 @@ import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.service.timeline.ha
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.service.timeline.handlers.TimelineTrancheCalculationHandler
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.service.timeline.handlers.TimelineTrancheThreeCalculationHandler
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.service.timeline.handlers.TimelineUalAdjustmentCalculationHandler
-import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.util.isAfterOrEqualTo
 import java.time.LocalDate
 
 @Service
@@ -51,6 +45,7 @@ class BookingTimelineService(
   private val timelineUalAdjustmentCalculationHandler: TimelineUalAdjustmentCalculationHandler,
   private val timelineExternalReleaseMovementCalculationHandler: TimelineExternalReleaseMovementCalculationHandler,
   private val timelineExternalAdmissionMovementCalculationHandler: TimelineExternalAdmissionMovementCalculationHandler,
+  private val timelineAdjustmentService: TimelineAdjustmentService,
   private val featureToggles: FeatureToggles,
 ) {
 
@@ -133,7 +128,7 @@ class BookingTimelineService(
         )
 
         if (featureToggles.adjustmentsAfterTrancheEnabled) {
-          latestCalculation = applyTrancheAdjustmentLogic(latestCalculation, adjustments, trancheAndCommencement)
+          latestCalculation = timelineAdjustmentService.applyTrancheAdjustmentLogic(latestCalculation, adjustments, trancheAndCommencement)
         }
       }
 
@@ -222,153 +217,6 @@ class BookingTimelineService(
       .sortedBy { it.date }
       .distinct()
       .groupBy { it.date }
-  }
-
-  private fun applyTrancheAdjustmentLogic(
-    calculation: CalculationResult,
-    adjustments: Adjustments,
-    trancheAndCommencement: Pair<SDSEarlyReleaseTranche, LocalDate?>,
-  ): CalculationResult {
-    if (trancheAndCommencement.first !in listOf(
-        SDSEarlyReleaseTranche.TRANCHE_1,
-        SDSEarlyReleaseTranche.TRANCHE_2,
-      )
-    ) {
-      log.info("No adjustments to apply as tranche is ${calculation.sdsEarlyReleaseTranche}")
-      return calculation
-    }
-
-    return adjustReleaseDatesForTranche(calculation, adjustments, trancheAndCommencement.second!!)
-  }
-
-  private fun adjustReleaseDatesForTranche(
-    calculation: CalculationResult,
-    adjustments: Adjustments,
-    commencementDate: LocalDate,
-  ): CalculationResult {
-    val rulesToHandle = listOf(
-      CalculationRule.SDS_EARLY_RELEASE_ADJUSTED_TO_TRANCHE_1_COMMENCEMENT,
-      CalculationRule.SDS_EARLY_RELEASE_ADJUSTED_TO_TRANCHE_2_COMMENCEMENT,
-      CalculationRule.SDS_STANDARD_RELEASE_APPLIES,
-    )
-
-    val releaseDateTypesToAdjust = listOf(
-      ReleaseDateType.HDCED,
-      ReleaseDateType.ERSED,
-      ReleaseDateType.PED,
-    )
-
-    var updatedCalculation = calculation
-
-    for (rule in rulesToHandle) {
-      for (releaseDateType in releaseDateTypesToAdjust) {
-        updatedCalculation = adjustReleaseDateForType(
-          updatedCalculation,
-          adjustments,
-          commencementDate,
-          releaseDateType,
-          rule,
-        )
-      }
-    }
-
-    return updatedCalculation
-  }
-
-  private fun adjustReleaseDateForType(
-    calculation: CalculationResult,
-    adjustments: Adjustments,
-    commencementDate: LocalDate,
-    releaseDateType: ReleaseDateType,
-    rule: CalculationRule,
-  ): CalculationResult {
-    val applicableRules = calculation.breakdownByReleaseDateType[releaseDateType]?.rules
-
-    if (applicableRules?.contains(rule) == true) {
-      log.info("${releaseDateType.name} before adjustment: ${calculation.dates[releaseDateType]}")
-      val crd = calculation.dates[ReleaseDateType.CRD]!!
-
-      val numberOfDaysToAdjust = calculateAdjustmentDaysAfterCommencement(adjustments, commencementDate, crd)
-
-      if (numberOfDaysToAdjust != 0L) {
-        return applyAdjustment(
-          calculation,
-          releaseDateType,
-          commencementDate,
-          rule,
-          numberOfDaysToAdjust,
-        )
-      }
-    }
-
-    return calculation
-  }
-
-  private fun applyAdjustment(
-    calculation: CalculationResult,
-    releaseDateType: ReleaseDateType,
-    commencementDate: LocalDate,
-    rule: CalculationRule,
-    numberOfDaysToAdjust: Long,
-  ): CalculationResult {
-    var updatedCalculation = calculation.copy(
-      dates = calculation.dates.plus(
-        releaseDateType to calculation.dates[releaseDateType]!!.plusDays(numberOfDaysToAdjust),
-      ),
-    )
-    log.info("$releaseDateType after adjustment: ${updatedCalculation.dates[releaseDateType]}")
-
-    if (updatedCalculation.dates[releaseDateType]?.isBefore(commencementDate) == true &&
-      rule in listOf(
-        CalculationRule.SDS_EARLY_RELEASE_ADJUSTED_TO_TRANCHE_1_COMMENCEMENT,
-        CalculationRule.SDS_EARLY_RELEASE_ADJUSTED_TO_TRANCHE_2_COMMENCEMENT,
-      )
-    ) {
-      updatedCalculation = updatedCalculation.copy(
-        dates = updatedCalculation.dates.plus(
-          releaseDateType to commencementDate,
-        ),
-      )
-    }
-
-    return updateBreakdownWithTrancheAdjustedRule(updatedCalculation, releaseDateType)
-  }
-
-  private fun calculateAdjustmentDaysAfterCommencement(
-    adjustments: Adjustments,
-    commencementDate: LocalDate,
-    crd: LocalDate,
-  ): Long {
-    val ualDays = adjustments.getOrEmptyList(UNLAWFULLY_AT_LARGE)
-      .filter { it.appliesToSentencesFrom.isAfterOrEqualTo(commencementDate) && it.appliesToSentencesFrom.isBefore(crd) }
-      .sumOf { it.numberOfDays }
-
-    val adaDays = adjustments.getOrEmptyList(ADDITIONAL_DAYS_AWARDED)
-      .filter { it.appliesToSentencesFrom.isAfterOrEqualTo(commencementDate) && it.appliesToSentencesFrom.isBefore(crd) }
-      .sumOf { it.numberOfDays }
-
-    val radaDays = adjustments.getOrEmptyList(RESTORATION_OF_ADDITIONAL_DAYS_AWARDED)
-      .filter { it.appliesToSentencesFrom.isAfterOrEqualTo(commencementDate) && it.appliesToSentencesFrom.isBefore(crd) }
-      .sumOf { it.numberOfDays }
-
-    val totalAdjustmentsToApply = ualDays + adaDays - radaDays
-    log.info("Total adjustments to apply $totalAdjustmentsToApply for commencement date $commencementDate (UAL $ualDays + ADAs $adaDays + RADA $radaDays)")
-
-    return totalAdjustmentsToApply.toLong()
-  }
-
-  private fun updateBreakdownWithTrancheAdjustedRule(
-    calculation: CalculationResult,
-    releaseDateType: ReleaseDateType,
-  ): CalculationResult {
-    val existingBreakdown = calculation.breakdownByReleaseDateType[releaseDateType]
-    return calculation.copy(
-      breakdownByReleaseDateType = calculation.breakdownByReleaseDateType.plus(
-        releaseDateType to existingBreakdown!!.copy(
-          rules = existingBreakdown.rules.plus(ADJUSTED_AFTER_TRANCHE_COMMENCEMENT),
-        ),
-      ),
-    )
   }
 
   companion object {
